moved header files and made verifyCC static library
